enterprise-rag/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-multipart==0.0.9
supabase==2.7.0
openai==1.45.0
tiktoken==0.7.0
pypdf2==3.0.1
python-docx==1.1.2
pandas==2.2.0
openpyxl==3.1.5
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.3.0
langchain-text-splitters==0.2.4
nltk==3.9.0
anyio==4.4.0
aiofiles==24.1.0
pytest==8.3.0
pytest-asyncio==0.24.0
